updated paths, updated data in avail_locations
