+++ utils/available_locations/gold/city-list.js
@@ -1,0 +1,14 @@
+// https://www.goodreturns.in/gold-rates/
+
+const Indian_Gold_Cities = `
+[
+
+]
+`;
+
+const getIndian_Gold_Cities = function ()
+{
+  return Indian_Gold_Cities;
+}
+
+module.exports = getIndian_Gold_Cities;
