+++ utils/available_locations/gold/country-list.js
@@ -1,0 +1,72 @@
+// https://rates.goldenchennai.com/world/gold-rate/
+
+
+const Gold_Countries = `
+[
+  "Afghanistan",
+  "Algeria",
+  "Argentina",
+  "Australia",
+  "Austria",
+  "Bangladesh",
+  "Belgium",
+  "Bhutan",
+  "Brazil",
+  "Canada",
+  "Chile",
+  "China",
+  "Cuba",
+  "Denmark",
+  "Egypt",
+  "Finland",
+  "France",
+  "Germany",
+  "Greece",
+  "Hong Kong",
+  "Hungary",
+  "Indonesia",
+  "Iran",
+  "Ireland",
+  "Israel",
+  "Italy",
+  "Japan",
+  "Jordan",
+  "Maldives",
+  "Mauritius",
+  "Mexico",
+  "Morocco",
+  "Myanmar",
+  "Nepal",
+  "Netherlands",
+  "New Zealand",
+  "Nigeria",
+  "Norway",
+  "Pakistan",
+  "Philippines",
+  "Poland",
+  "Portugal",
+  "Romania",
+  "Russia",
+  "South Africa",
+  "South Korea",
+  "Spain",
+  "Sri Lanka",
+  "Sweden",
+  "Switzerland",
+  "Syria",
+  "Thailand",
+  "Turkey",
+  "UK",
+  "USA",
+  "Vietnam",
+  "Yemen"
+]
+`;
+
+
+const getGold_Countries = function ()
+{
+    return Gold_Countries;
+}
+
+module.exports = getGold_Countries;
