--- utils/available_locations/gold/gold-1.js
@@ -1,118 +1,0 @@
-// https://www.goodreturns.in/gold-rates/
-
-const Indian_Gold_Cities = `
-[
-  "Agra",
-  "Ahmedabad",
-  "Amaravati",
-  "Ambur",
-  "Amravati",
-  "Amritsar",
-  "Anantapur",
-  "Arcot",
-  "Ariyalur",
-  "Aurangabad",
-  "Bangalore",
-  "Bellary",
-  "Berhampur",
-  "Bhiwandi",
-  "Bhopal",
-  "Bhubaneswar",
-  "Chandigarh",
-  "Chennai",
-  "Coimbatore",
-  "Cuttack",
-  "Davanagere",
-  "Delhi",
-  "Dharmapuri",
-  "Dindigul",
-  "Erode",
-  "Ghaziabad",
-  "Goa",
-  "Guntur",
-  "Gurgaon",
-  "Hyderabad",
-  "Indore",
-  "Jaipur",
-  "Jayankondam",
-  "Kakinada",
-  "Kallakurichi",
-  "Kanchipuram",
-  "Kanpur",
-  "Karaikudi",
-  "Karur",
-  "Khammam",
-  "Kochi",
-  "Kodaikanal",
-  "Kolhapur",
-  "Kolkata",
-  "Kovilpatti",
-  "Krishnagiri",
-  "Kumbakonam",
-  "Kanyakumari",
-  "Latur",
-  "Lucknow",
-  "Ludhiana",
-  "Madurai",
-  "Mangalore",
-  "Mohali",
-  "Mumbai",
-  "Mysore",
-  "Nagapattinam",
-  "Nagpur",
-  "Nagercoil",
-  "Nashik",
-  "Nellore",
-  "Nizamabad",
-  "Noida",
-  "Ooty",
-  "Palani",
-  "Paramakudi",
-  "Patna",
-  "Perambalur",
-  "Pune",
-  "Pudukkottai",
-  "Ramanathapuram",
-  "Rameswaram",
-  "Rajkot",
-  "Rourkela",
-  "Salem",
-  "Sambalpur",
-  "Sivagangai",
-  "Solapur",
-  "Surat",
-  "Thanjavur",
-  "Thane",
-  "Theni",
-  "Tirupati",
-  "Tirupur",
-  "Tirunelvelli",
-  "Tiruvarur",
-  "Tiruvannamalai",
-  "Trichy",
-  "Trivandrum",
-  "Tuticorin",
-  "Vadodara",
-  "Vasai-Virar",
-  "Vellore",
-  "Vijayawada",
-  "Villupuram",
-  "Visakhapatnam",
-  "Virudhunagar",
-  "Warangal",
-  "Kolkata",
-  "Kochi",
-  "Kodaikanal",
-  "Kumbakonam",
-  "Ludhiana",
-  "Madurai",
-  "Mangalore"
-]
-`;
-
-const getIndian_Gold_Cities = function ()
-{
-  return Indian_Gold_Cities;
-}
-
-module.exports = getIndian_Gold_Cities;
