--- utils/available_locations/gold/gold-2.js
@@ -1,47 +1,0 @@
-// https://pricegold.net/today/
-
-const Gold_Countries = `
-[
-    {"code": "DZ", "name": "Algeria"},
-    {"code": "AU", "name": "Australia"},
-    {"code": "BH", "name": "Bahrain"},
-    {"code": "BR", "name": "Brazil"},
-    {"code": "CA", "name": "Canada"},
-    {"code": "CN", "name": "China"},
-    {"code": "EG", "name": "Egypt"},
-    {"code": "EU", "name": "Europe"},
-    {"code": "GH", "name": "Ghana"},
-    {"code": "IN", "name": "India"},
-    {"code": "ID", "name": "Indonesia"},
-    {"code": "IR", "name": "Iran"},
-    {"code": "IQ", "name": "Iraq"},
-    {"code": "JO", "name": "Jordan"},
-    {"code": "KW", "name": "Kuwait"},
-    {"code": "LB", "name": "Lebanon"},
-    {"code": "LY", "name": "Libya"},
-    {"code": "MX", "name": "Mexico"},
-    {"code": "MA", "name": "Morocco"},
-    {"code": "OM", "name": "Oman"},
-    {"code": "PK", "name": "Pakistan"},
-    {"code": "QA", "name": "Qatar"},
-    {"code": "RU", "name": "Russia"},
-    {"code": "SA", "name": "Saudi Arabia"},
-    {"code": "SG", "name": "Singapore"},
-    {"code": "ZA", "name": "South Africa"},
-    {"code": "KR", "name": "South Korea"},
-    {"code": "TR", "name": "Turkey"},
-    {"code": "AE", "name": "United Arab Emirates"},
-    {"code": "GB", "name": "United Kingdom"},
-    {"code": "US", "name": "United States"},
-    {"code": "UZ", "name": "Uzbekistan"},
-    {"code": "YE", "name": "Yemen"}
-  ]
-`;
-
-
-const getGold_Countries = function ()
-{
-    return Gold_Countries;
-}
-
-module.exports = getGold_Countries;
