+++ utils/available_locations/silver/city-list.js
@@ -1,0 +1,14 @@
+// https://www.goodreturns.in/silver-rates/
+
+const Indian_Silver_Cities = `
+[
+ 
+]
+`;
+
+const getIndian_Silver_Cities = function ()
+{
+  return Indian_Silver_Cities;
+}
+
+module.exports = getIndian_Silver_Cities;
